Moving stepper motor and pump controls into header
