+++ src/Main.cpp
@@ -10,7 +10,6 @@
 #include <WebServer.h>
 #include <DNSServer.h>
 #include <Update.h>
-#include <ESP_FlexyStepper.h>
 #include <Wire.h>
 
 //all file includes
@@ -18,43 +17,10 @@
 #include "FlowSensor.hpp"
 #include "Pump.hpp"
 #include "BioreactorVaribiles.hpp"
-
-// Set up Pump controller
-const int MODBUS_RX2 = 17;
-const int MODBUS_TX2 = 5;
-const int MODBUS_DE = 19;
-const int MODBUS_RE = 19;
-const int MODBUS_ENABLE = 19; // automatically set to high when writing, low otherwise to receive
-const int PUMP_ADDRESS = 0xEF; // Modbus address of pump controller
-const int MODBUS_TIMEOUT = 500; // timeout in ms for Modbus command responses
-
-void preTransmission()
-{
-  digitalWrite(MODBUS_RE, 1);
-  digitalWrite(MODBUS_DE, 1);
-}
-
-void postTransmission()
-{
-  digitalWrite(MODBUS_RE, 0);
-  digitalWrite(MODBUS_DE, 0);
-}
+#include "StepperMotor.hpp"
 
 //Set up Flow Sensor
 SensirionLF flowSensor(SLF3X_SCALE_FACTOR_FLOW, SLF3X_SCALE_FACTOR_TEMP, SLF3X_I2C_ADDRESS);
-
-//Pins for Stepper Motor
-int HIGH_MOTOR_DIRPIN = 27;
-int HIGH_MOTOR_STEPPIN = 26;
-int HIGH_MOTOR_ENAPIN = 25;
-
-//Set up Stepper Motor varibiles
-ESP_FlexyStepper stepper;
-const float STEPS_PER_REV = 200;
-const float DISTANCE_PER_REV = 1;
-const float MAX_SPEED = 1000;
-const float ACCELERATION = 1000;
-const float MOVE_DISTANCE = 5; 
 
 //Start Running
 void setup() {
@@ -83,20 +49,11 @@ void setup() {
   pumpOn = checkStatus();
   Serial.printf("Pump is: %d\n", pumpOn);
   
-  //Set up Flow Sensor
+  //Set up Flow Sensor and Stepper Motor
   flowSensorSetup(flowSensor); //Function in FlowSensor.hpp
-
   delay(100);
+  stepperSetup(stepper); //Function in StepperMotor.hpp
 
-  //Setup Stepper Motor Paramaters
-  stepper.connectToPins(HIGH_MOTOR_STEPPIN, HIGH_MOTOR_DIRPIN);
-
-  stepper.setStepsPerMillimeter(STEPS_PER_REV / DISTANCE_PER_REV);
-  stepper.setSpeedInStepsPerSecond(MAX_SPEED);
-  stepper.setAccelerationInStepsPerSecondPerSecond(ACCELERATION);
-
-  //stepper.setCurrentPositionInMillimeters(0.0);
-  stepper.startAsService(1);
 }
 
 void loop() {
